<refactor> : ApiResponse 부모 클래스 생성
Response DTO의 반복되는 부분이 많아 해당 부모 클래스를 만들어 상속으로 간결하게 리팩토링
