+++ src/main/java/com/example/demo/domain/dto/response/ApiResponse.java
@@ -1,0 +1,28 @@
+package com.example.demo.domain.dto.response;
+
+public class ApiResponse {
+
+    private String message;
+    private int status;
+
+    public ApiResponse(String message, int status) {
+        this.message = message;
+        this.status = status;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+
+    public int getStatus() {
+        return status;
+    }
+
+    public void setStatus(int status) {
+        this.status = status;
+    }
+}
