+++ src/main/java/com/example/demo/domain/dto/response/LoginResponse.java
@@ -1,30 +1,8 @@
 package com.example.demo.domain.dto.response;
 
-public class LoginResponse {
+public class LoginResponse extends ApiResponse{
 
-    private String message;
-    private int status;
-
-    // Constructor
     public LoginResponse(String message, int status) {
-        this.message = message;
-        this.status = status;
-    }
-
-    // Getters and Setters
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-
-    public int getStatus() {
-        return status;
-    }
-
-    public void setStatus(int status) {
-        this.status = status;
+        super(message,status);
     }
 }
