+++ src/main/java/com/example/demo/domain/dto/response/RegisterResponse.java
@@ -1,30 +1,8 @@
 package com.example.demo.domain.dto.response;
 
-public class RegisterResponse {
+public class RegisterResponse extends ApiResponse{
 
-    private String message;
-    private int status;
-
-    // Constructor
     public RegisterResponse(String message, int status) {
-        this.message = message;
-        this.status = status;
-    }
-
-    // Getters and Setters
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-
-    public int getStatus() {
-        return status;
-    }
-
-    public void setStatus(int status) {
-        this.status = status;
+        super(message,status);
     }
 }
